// src/api/generated/models/playerResource.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Laravel
 * OpenAPI spec version: 0.0.1
 */
import type { UserWithClubResource } from './userWithClubResource.ts';

export interface PlayerResource {
  id: string;
  name: string;
  overall: number;
  position: string;
  nationality: string;
  /** @nullable */
  category: string | null;
  /** @nullable */
  image_url: null;
  /** @nullable */
  salary: string | null;
  /** @nullable */
  passe: string | null;
  owner: UserWithClubResource | null;
  created_at: string;
  updated_at: string;
}